Move analytics charts component
