+++ web/components/charts/stats.tsx
@@ -1,15 +1,12 @@
 import { useMemo } from 'react'
 import { scaleTime, scaleLinear } from 'd3-scale'
-import { min, max } from 'lodash'
+import { max } from 'lodash'
 import dayjs from 'dayjs'
 
 import { formatPercent } from 'common/util/format'
 import { Row } from '../layout/row'
-import {
-  HistoryPoint,
-  SingleValueHistoryChart,
-} from 'web/components/charts/generic-charts'
-import { TooltipProps, MARGIN_X, MARGIN_Y } from 'web/components/charts/helpers'
+import { HistoryPoint, SingleValueHistoryChart } from './generic-charts'
+import { TooltipProps, MARGIN_X, MARGIN_Y } from './helpers'
 import { SizedContainer } from 'web/components/sized-container'
 
 const getPoints = (startDate: number, dailyValues: number[]) => {
